requirements.txt
flask==3.0.2
google-generativeai>=0.3.2
ShopifyAPI==12.0.0
numpy==1.26.4
python-dotenv==1.0.1
requests>=2.32.3
flask-cors==4.0.0
